refactor: split config/file/main.rs into focused modules
Convert main.rs into a directory module with mod.rs as a pure
aggregator. Production code splits along natural seams: the Config
struct/types in config.rs, impl Config in lifecycle.rs, and the
read_config / parse_config / deprecated-key migration pipeline in
parse.rs. Each module owns its tests inline.

Public API is unchanged: pub use config::* and pub use parse::* in
mod.rs keep crate::config::file::Config, parse_config, etc. accessible
at the same path as before.
